Split apart serializers
