+++ addon/attributes/array.js
@@ -54,7 +54,7 @@ export default function array(type, options) {
 
   // Use computed with a dependency on hasDirtyAttributes which changes on rollback
   // This ensures the computed property is re-evaluated when dirty state changes
-  return computed('hasDirtyAttributes', 'currentState', {
+  return computed('currentState', 'hasDirtyAttributes', 'store.cache', {
     get(key) {
       if (this.isDestroying || this.isDestroyed) {
         return null;
